refactor: 관리자 기능 이관
