+++ src/main/java/com/est/cranejob/user/controller/AdminUserController.java
@@ -1,5 +1,6 @@
 package com.est.cranejob.user.controller;
 
+import com.est.cranejob.announcement.dto.AdminUpdateUserRequest;
 import com.est.cranejob.user.dto.request.CreateAdminUserRequest;
 import com.est.cranejob.user.dto.request.UpdateAdminUserRequest;
 import com.est.cranejob.user.dto.response.UserResponse;
@@ -7,7 +8,13 @@ import com.est.cranejob.user.service.AdminUserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -17,8 +24,11 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequiredArgsConstructor
@@ -100,5 +110,47 @@ public class AdminUserController {
 		}
 
 		return "redirect:/";
+	}
+
+	// 공지사항에 대한 CRUD 로직 들어가야함
+	@GetMapping("/admin/users")
+	public String userList(@RequestParam("page") Optional<Integer> page, @RequestParam("size") Optional<Integer> size, @RequestParam("keyword") Optional<String> keyword, Model model) {
+
+		int currentPage = page.orElse(1);
+		int pageSize = size.orElse(10);
+		String searchKeyword = keyword.orElse("");
+
+		Page<UserResponse> userResponses = adminUserService.getPaginatedUsers(
+			PageRequest.of(currentPage - 1, pageSize), searchKeyword);
+
+		model.addAttribute("userResponses", userResponses);
+
+		int totalPages = userResponses.getTotalPages();
+
+		if (totalPages > 0) {
+			List<Integer> pageNumbers = IntStream.rangeClosed(1, totalPages)
+				.boxed()
+				.collect(Collectors.toList());
+			model.addAttribute("pageNumbers", pageNumbers);
+		}
+
+		return "/admin/user-list";
+	}
+
+	@GetMapping("/admin/users/{username}")
+	public String userDetails(@PathVariable String username, Model model) {
+
+		UserResponse userResponse = adminUserService.findUserByUsername(username);
+
+		model.addAttribute("adminUpdateUserRequest", AdminUpdateUserRequest.toResponseDto(userResponse));
+
+		return "/admin/user-info";
+	}
+
+	@PostMapping("/admin/users/edit")
+	public String updateUserRole(@ModelAttribute AdminUpdateUserRequest adminUpdateUserRequest) {
+		adminUserService.updateUserRole(adminUpdateUserRequest.getUsername(), adminUpdateUserRequest.getRole(), adminUpdateUserRequest.getUserStatus());
+
+		return "redirect:/admin/users";
 	}
 }
