+++ src/main/resources/templates/admin/user-list.html
@@ -31,48 +31,6 @@
 <div class="layout">
 
 <h2 align="center">유저 내역 리스트</h2>
-<nav class="navbar navbar-expand-lg bg-light">
-	
-	<div class="container-fluid">
-		<a class="navbar-brand" href="/post/list">게시판</a>
-		<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-			<span class="navbar-toggler-icon"></span>
-		</button>
-		
-		<div class="collapse navbar-collapse" id="navbarSupportedContent">
-			<ul class="navbar-nav ms-auto">
-				
-				<!-- 세션이 존재하지 않을 경우 로그인, 회원가입 버튼만 활성화 -->
-				<li class="nav-item">
-					<a class="nav-link" href="/user/login" sec:authorize="isAnonymous()">Login</a>
-				</li>
-				<li class="nav-item">
-					<a class="nav-link" href="/user/signup" sec:authorize="isAnonymous()">Join</a>
-				</li>
-				
-				<!-- 세션이 존재할 경우 로그아웃, 개인정보수정 버튼으로 대체 -->
-				<li class="nav-item" sec:authorize="isAuthenticated()">
-					<a class="nav-link" th:text="|로그인: ${#authentication.principal.nickname}|">로그인 사용자 이름</a>
-				</li>
-				
-				<li class="nav-item" sec:authorize="isAuthenticated()">
-					<a class="nav-link" href="/user/logout">Logout</a>
-				</li>
-				
-				<li class="nav-item" sec:authorize="hasRole('ROLE_USER')">
-					<a class="nav-link" href="/user/edit">내 정보 수정</a>
-				</li>
-				
-				<li class="nav-item" sec:authorize="hasRole('ROLE_ADMIN')">
-					<a class="nav-link" href="/admin/edit">내 정보 수정</a>
-				</li>
-				<li class="nav-item" sec:authorize="hasRole('ROLE_ADMIN')">
-					<a class="nav-link" href="/admin/user-list">회원 정보 수정</a>
-				</li>
-			</ul>
-		</div>
-	</div>
-</nav>
 <div class="layout">
 	<div style="width: 300px" class="input-group mx-auto mb-3">
 		<form class="row gx-2" th:action="@{/admin/users}" method="get">
