+++ src/main/resources/templates/layout/header.html
@@ -93,7 +93,7 @@
                     <a class="nav-links" href="/admin/edit">내 정보 수정</a>
                 </li>
                 <li class="nav-items" sec:authorize="hasRole('ROLE_ADMIN')">
-                    <a class="nav-links" href="/admin/user-list">회원 정보 수정</a>
+                    <a class="nav-links" href="/admin/users">회원 정보 수정</a>
                 </li>
             </ul>
         </div>
